todos/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
dj_rest_auth==1.1.1
dj-rest-auth[with_social]
psycopg2-binary===2.8.6